+++ src/fuzzy_asteroids/game.py
@@ -88,7 +88,6 @@ class AsteroidGame(arcade.Window):
 
         # Set up the game instance
         self.game_over = None
-        self.controller_name = None
 
         # Evaluation analytics
         self.score = None
@@ -161,6 +160,7 @@ class AsteroidGame(arcade.Window):
             self.set_size(self.scenario.game_map.width, self.scenario.game_map.height)
 
         self._print_terminal("**********************************************************")
+        self._print_terminal(f"Controller: {self.controller.name if hasattr(self, 'controller') else ''}")
         self._print_terminal(f"Scenario: {self.scenario.name}")
         self._print_terminal(f"- - - - - - - - - - - - - - - - - - - - - - - - - - - - -")